Override default list layout
The default list layout for the `terminal` theme only lists pages with
the content type specified in config.toml (or "posts" if none is
specified). This layout changes that to list all pages of the current
type unless the home page is being rendered, in which case it will list
the pages of the content type specified in config.toml.